move the sendRegistrationData function to parent components as a callback to improve reusability
