+++ src/pages/member/EventDetails.js
@@ -10,6 +10,7 @@ import QuickRegister from "./QuickRegister";
 import ArrowBackIcon from "@material-ui/icons/ArrowBack";
 import Typography from "@material-ui/core/Typography";
 import { REGISTRATION_STATUS } from "../../constants/Constants";
+import { fetchBackend } from '../../utils'
 
 const useStyles = makeStyles(theme => ({
   layout: {
@@ -95,6 +96,51 @@ const EventDetails = props => {
     console.log('eventRegistrationStatus:', eventRegistrationStatus);
   }
 
+  let settingRegistrationData = false;
+
+  const sendRegistrationDataCallback = async (id, eventID, isRegister, isFirstTime, heardFrom) => {
+    if (settingRegistrationData === true) {
+      return Promise.resolve("in_progress");
+    }
+    settingRegistrationData = true;
+    let registrationStatus = "";
+    let method = "";
+    let path = "";
+    if (isRegister) {
+      registrationStatus = REGISTRATION_STATUS.REGISTERED;
+    } else {
+      registrationStatus = REGISTRATION_STATUS.CANCELLED;
+    }
+    let body = {
+      eventID: eventID,
+      registrationStatus: registrationStatus
+    };
+    if (isFirstTime) {
+      body["id"] = id;
+      method = "POST";
+      path = "/registrations";
+      if(heardFrom) {
+        body["heardFrom"] = heardFrom;
+      }
+    } else {
+      method = "PUT";
+      path = `/registrations/${id}`;
+    }
+    try {
+      await fetchBackend(path, method, body);
+      settingRegistrationData = false;
+      let responesMsg = "";
+      isRegister
+        ? (responesMsg = "registration")
+        : (responesMsg = "unregistration");
+      responesMsg += " succeed";
+      return Promise.resolve(responesMsg);
+    } catch (error) {
+      settingRegistrationData = false;
+      return Promise.reject(error);
+    }
+  };
+
   return event ? (
     <React.Fragment>
       <Helmet>
@@ -116,6 +162,7 @@ const EventDetails = props => {
             eventRegistrationStatus={eventRegistrationStatus}
             handleRegisterStateChangedCallback={handleRegisterStateChangedCallback}
             handleRegisterClickedCallback={handleRegisterClickedCallback}
+            sendRegistrationDataCallback={sendRegistrationDataCallback}
           />
         ) : (
           <EventDescription
@@ -125,6 +172,7 @@ const EventDetails = props => {
             eventRegistrationStatus={eventRegistrationStatus}
             handleRegisterStateChangedCallback={handleRegisterStateChangedCallback}
             handleRegisterClickedCallback={handleRegisterClickedCallback}
+            sendRegistrationDataCallback={sendRegistrationDataCallback}
           />
         )}
       </div>
